move the items into the correct directory
